Added ability to ride horses.
Extremely hacky approach was taken to fix an issue where horses being
ridden by a villager would not move. When a villager mounts a horse, the
server and all clients are told that it doesn't have a saddle, which
allows it to move but also removes the saddle from its model. To remedy
this, the horse model is replaced by an extension model that only
renders when ridden by a villager. It forces the saddle to appear and
animate only when a villager is riding the horse.