Fix: Sleeping logic simplified, fixing home point issues, suffocation, etc. Fixes #491, #485, #484, #341
Sleeping point now set to the foot of the bed, not the head of the bed.
Villagers actually placed at that location ran the risk of being caught
within an adjacent wall and suffocating.

Some villagers also did not register that a home point was invalid and
teleported anyway. This was fixed.

Villagers will no longer walk to a home point.