Fix targeting system to include slimes
Slimes do not extend EntityMob and so is the MagmaCubes, yet they are mobs since they implement IMob. By changing it to IMob, you give the chance for other mods to implement targets that are not EntityMob and will still be attack-able. This change will include shulkers, ghasts, slimes(Mgama cubes too) and dragons.